Moved tests from 'test_graphstate' into test_unit_conversion and test_code_generation (which is what they test).
