+++ test/test_unit_conversion.py
@@ -2,9 +2,67 @@
 Tests for unit conversion of inputs and defines.
 For further unit conversion tests see test_graphstate.py
 """
+import os
 import pytest
 
 import fc
+import fc.test_support
+
+
+def test_graphstate_time_conversion():
+    """ Tests the graph state protocol in a model requiring time units conversion. """
+
+    # Create protocol
+    proto = fc.Protocol(os.path.join('protocols', 'GraphState.txt'))
+
+    # Set model (generates & compiles model)
+    model_name = 'difrancesco_noble_model_1985'  # has time in seconds, not milliseconds
+    proto.set_model(os.path.join('test', 'models', model_name + '.cellml'))
+
+    # Run protocol
+    proto.set_output_folder('test_graphstate_time_conversion')
+    proto.run()
+    # Some test assertions are within the protocol itself
+
+    # Check output exists
+    assert os.path.exists(os.path.join(proto.output_folder.path, 'output.h5'))
+
+    # Check output is correct
+    assert fc.test_support.check_results(
+        proto,
+        {'state': 2},   # Name and dimension of output to check
+        os.path.join('test', 'data', 'historic', model_name, 'GraphState'),
+        rel_tol=0.005,
+        abs_tol=2.5e-4
+    )
+
+
+def test_graphstate_voltage_conversion():
+    """ Tests the graph state protocol in a model requiring state variable units conversion. """
+
+    # Create protocol
+    proto = fc.Protocol(os.path.join('protocols', 'GraphState.txt'))
+
+    # Set model (generates & compiles model)
+    model_name = 'paci_hyttinen_aaltosetala_severi_ventricularVersion'  # has voltage in volt, not millivolt
+    proto.set_model(os.path.join('test', 'models', model_name + '.cellml'))
+
+    # Run protocol
+    proto.set_output_folder('test_graphstate_voltage_conversion')
+    proto.run()
+    # Some test assertions are within the protocol itself
+
+    # Check output exists
+    assert os.path.exists(os.path.join(proto.output_folder.path, 'output.h5'))
+
+    # Check output is correct
+    assert fc.test_support.check_results(
+        proto,
+        {'state': 2},   # Name and dimension of output to check
+        os.path.join('test', 'data', 'historic', model_name, 'GraphState'),
+        rel_tol=0.005,
+        abs_tol=2.5e-4
+    )
 
 
 def test_unit_conversion_inputs_initial_values():
